tools/requirements.txt
pyserial
matplotlib
numpy
pytest
